[Fix #1491] PdrDashboard v0.4 - Split utils.py into concerns (#1493)
* First splitup.
* Adds database getter.
* Display tooltips from file.
