+++ pdr_backend/pdr_dashboard/callbacks/callbacks_home.py
@@ -6,15 +6,13 @@ from pdr_backend.pdr_dashboard.dash_components.plots import (
     get_figures_and_metrics,
 )
 
-from pdr_backend.pdr_dashboard.dash_components.util import (
+from pdr_backend.pdr_dashboard.util.data import (
     filter_objects_by_field,
     get_date_period_text,
-    get_feed_ids_based_on_predictoors_from_db,
-    get_payouts_from_db,
     get_start_date_from_period,
     select_or_clear_all_by_table,
-    calculate_tx_gas_fee_cost_in_OCEAN,
 )
+from pdr_backend.pdr_dashboard.util.prices import calculate_tx_gas_fee_cost_in_OCEAN
 from pdr_backend.pdr_dashboard.dash_components.view_elements import (
     get_graph,
 )
@@ -68,11 +66,10 @@ def get_callbacks_home(app):
                 if int(date_period) > 0
                 else 0
             )
-            payouts = get_payouts_from_db(
+            payouts = app.db_getter.payouts(
                 [row["contract"] for row in selected_feeds],
                 predictoors_addrs,
                 start_date,
-                app.lake_dir,
             )
 
         # get fee estimate
@@ -195,8 +192,7 @@ def get_callbacks_home(app):
 
         # filter feeds by payouts from selected predictoors
         if predictoor_feeds_only and (len(predictoors_addrs) > 0):
-            feed_ids = get_feed_ids_based_on_predictoors_from_db(
-                app.lake_dir,
+            feed_ids = app.db_getter.feed_ids_based_on_predictoors(
                 predictoors_addrs,
             )
             filtered_data = [
