+++ pdr_backend/pdr_dashboard/predictoor_dash.py
@@ -10,10 +10,9 @@ from pdr_backend.pdr_dashboard.callbacks.callbacks_home import (
 from pdr_backend.pdr_dashboard.callbacks.callbacks_common import (
     get_callbacks_common,
 )
-from pdr_backend.pdr_dashboard.dash_components.util import (
-    get_feeds_data_from_db,
-    get_predictoors_data_from_payouts,
-    get_user_payouts_stats_from_db,
+from pdr_backend.pdr_dashboard.util.db import DBGetter
+from pdr_backend.pdr_dashboard.util.data import get_predictoors_data_from_payouts
+from pdr_backend.pdr_dashboard.util.prices import (
     fetch_token_prices,
 )
 from pdr_backend.pdr_dashboard.dash_components.view_elements import (
@@ -46,12 +45,12 @@ def predictoor_dash(ppss: PPSS, debug_mode: bool):
 @enforce_types
 def setup_app(app, ppss: PPSS):
     app.web3_pp = ppss.web3_pp
-    app.lake_dir = ppss.lake_ss.lake_dir
+    app.db_getter = DBGetter(ppss.lake_ss.lake_dir)
     app.network_name = ppss.web3_pp.network
 
-    app.feeds_data = get_feeds_data_from_db(ppss.lake_ss.lake_dir)
+    app.feeds_data = app.db_getter.feeds_data()
     app.predictoors_data = get_predictoors_data_from_payouts(
-        get_user_payouts_stats_from_db(ppss.lake_ss.lake_dir)
+        app.db_getter.payouts_stats()
     )
     app.favourite_addresses = ppss.predictoor_ss.my_addresses
     app.layout = get_layout()
